Implementação JWT authentication

+++ src/main/java/school/sptech/cursos/service/UsuarioService.java
@@ -1,8 +1,17 @@
 package school.sptech.cursos.service;
 
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
+import school.sptech.cursos.Config.GerenciadorTokenJwt;
 import school.sptech.cursos.DTO.Usuario.UsuarioRequest;
 import school.sptech.cursos.DTO.Usuario.UsuarioResponse;
+import school.sptech.cursos.DTO.Usuario.UsuarioToken;
 import school.sptech.cursos.model.Usuario;
 import school.sptech.cursos.repository.IUsuarioRepository;
 
@@ -12,6 +21,15 @@ import java.util.List;
 @Service
 public class UsuarioService {
     private final IUsuarioRepository repository;
+
+    @Autowired
+    private PasswordEncoder passwordEncoder;
+
+    @Autowired
+    private GerenciadorTokenJwt gerenciadorTokenJwt;
+
+    @Autowired
+    private AuthenticationManager authenticationManager;
 
     public UsuarioService(IUsuarioRepository repository) {
         this.repository = repository;
@@ -45,9 +63,10 @@ public class UsuarioService {
         if (repository.existsByTelefone(request.getTelefone())) {
             throw new RuntimeException("Telefone já cadastrado");
         }
+        String senhaCriptografada = passwordEncoder.encode(request.getSenha());
         usuario.setNome(request.getNome());
         usuario.setEmail(request.getEmail());
-        usuario.setSenha(request.getSenha());
+        usuario.setSenha(senhaCriptografada);
         usuario.setTelefone(request.getTelefone());
         usuario.setTipoUsuario(request.getTipoUsuario());
         usuario = repository.save(usuario);
@@ -77,23 +96,26 @@ public class UsuarioService {
 
     }
 
-    public UsuarioResponse buscarPorEmailESenha(String email, String senha) {
+    public UsuarioToken autenticar(String email, String senha) {
+        var credentials = new UsernamePasswordAuthenticationToken(email, senha);
+        Authentication authentication = authenticationManager.authenticate(credentials);
+
         Usuario usuario = repository.findByEmail(email)
-                .orElseThrow(() -> new RuntimeException("Email ou senha inválidos"));
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Email não cadastrado"));
 
-        if (!senha.equals(usuario.getSenha())) {
-            throw new RuntimeException("Email ou senha inválidos");
-        }
+        String token = gerenciadorTokenJwt.generateToken(authentication);
 
-        UsuarioResponse response = new UsuarioResponse();
+        UsuarioToken response = new UsuarioToken();
         response.setId(usuario.getId());
         response.setNome(usuario.getNome());
         response.setEmail(usuario.getEmail());
         response.setTelefone(usuario.getTelefone());
         response.setTipoUsuario(usuario.getTipoUsuario());
+        response.setToken(token);
 
         return response;
     }
+
 
     public UsuarioResponse atualizar(Long id, UsuarioRequest request) {
         Usuario usuario = repository.findById(id)
@@ -111,7 +133,7 @@ public class UsuarioService {
 
         usuario.setNome(request.getNome());
         usuario.setEmail(request.getEmail());
-        usuario.setSenha(request.getSenha()); // tem qu cripitografar -> passar pr ocaiobaaa ou henry
+        usuario.setSenha(passwordEncoder.encode(request.getSenha()));
         usuario.setTelefone(request.getTelefone());
         usuario.setTipoUsuario(request.getTipoUsuario());
 
@@ -128,6 +150,9 @@ public class UsuarioService {
     }
 
     public void deletarPorID(Long id) {
+        if (!repository.existsById(id)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Usuário não encontrado");
+        }
         repository.deleteById(id);
     }
 }
